% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perufauna-package.R
\docType{package}
\name{perufauna-package}
\alias{perufauna}
\alias{perufauna-package}
\title{perufauna: Easily Load and Harmonize Peruvian Fauna and Biodiversity Packages}
\description{
The \code{perufauna} metapackage streamlines the installation, loading, and
integration of specialized R packages focused on Peruvian fauna and biodiversity:
\itemize{
\item \code{avesperu}: Taxonomic backbone and search tools for the birds of Peru (UNOP checklist).
\item \code{perumammals}: Taxonomic backbone and validation for mammals of Peru (Pacheco et al. 2021).
\item \code{citesperu}: Status and verification of species in CITES Peru appendices (MINAM).
\item \code{perufaunads004}: Threatened fauna categorization according to D.S. N° 004-2014-MINAGRI and SERFOR Red Book.
}

In addition to orchestrated attachment, \code{perufauna} provides integrated cross-referencing
functions such as \code{\link{pf_match}} and \code{\link{pf_status}} to analyze species across
all taxonomic and regulatory sources in a single call.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/perufauna}
  \item \url{https://paulesantos.github.io/perufauna/}
  \item Report bugs at \url{https://github.com/PaulESantos/perufauna/issues}
}

}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Paul E. Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID}) [copyright holder]
}

}
\keyword{internal}
