% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{perufauna_attach}
\alias{perufauna_attach}
\title{Attach Core Packages in perufauna}
\usage{
perufauna_attach()
}
\value{
Invisible character vector of newly attached packages.
}
\description{
Attaches all core member packages of the perufauna ecosystem
(\code{avesperu}, \code{perumammals}, \code{citesperu}, \code{perufaunads004})
that are not yet loaded onto the search path.
}
\examples{
\donttest{
perufauna_attach()
}
}
