% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{perufauna_deps}
\alias{perufauna_deps}
\title{List all perufauna dependencies and versions}
\usage{
perufauna_deps()
}
\value{
A \code{tibble::tibble} with package names, installed status, and local versions.
}
\description{
List all perufauna dependencies and versions
}
\examples{
perufauna_deps()
}
