% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{perufauna_logo}
\alias{perufauna_logo}
\title{The perufauna Logo in ASCII}
\usage{
perufauna_logo()
}
\value{
An object of class \code{perufauna_logo}.
}
\description{
Displays the stylized perufauna ASCII banner.
}
\examples{
perufauna_logo()
}
