% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perufauna_packages}
\alias{perufauna_packages}
\title{List all core packages in perufauna}
\usage{
perufauna_packages(include_self = FALSE)
}
\arguments{
\item{include_self}{Include perufauna in the list? Default is \code{FALSE}.}
}
\value{
Character vector of package names.
}
\description{
Lists the constituent member packages of the perufauna ecosystem.
}
\examples{
perufauna_packages()
}
