% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{perufauna_sitrep}
\alias{perufauna_sitrep}
\title{Situation Report on perufauna and Core Biodiversity Packages}
\usage{
perufauna_sitrep()
}
\description{
Prints a diagnostic overview of the current R session, the version of
\code{perufauna}, and the installation status and version of each core
member package (\code{avesperu}, \code{perumammals}, \code{citesperu}, \code{perufaunads004}).
}
\examples{
\donttest{
perufauna_sitrep()
}
}
