% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pf_status.R
\name{pf_status}
\alias{pf_status}
\title{Quick Status Summary for Peruvian Fauna}
\usage{
pf_status(splist, ...)
}
\arguments{
\item{splist}{Character vector of species names.}

\item{...}{Arguments forwarded to \code{\link{pf_match}}.}
}
\value{
A character vector of species status summaries.
}
\description{
Provides an immediate formatted textual summary of Peru occurrence,
taxonomic group, CITES appendix, and national threat status for each species.
}
\examples{
\donttest{
pf_status(c("Panthera onca", "Vultur gryphus", "Tremarctos ornatus"))
}
}
